+++ settings.gradle
@@ -12,6 +12,7 @@ include 'spring-instrument-tomcat'
 include 'spring-jdbc'
 include 'spring-jms'
 include 'spring-orm'
+include 'spring-orm-hibernate4'
 include 'spring-oxm'
 include 'spring-struts'
 include 'spring-test'
