+++ spring-orm-hibernate4/src/main/java/org/springframework/orm/hibernate4/HibernateTransactionManager.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2011 the original author or authors.
+ * Copyright 2002-2012 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,6 @@
 
 package org.springframework.orm.hibernate4;
 
-import java.lang.reflect.Method;
 import java.sql.Connection;
 import javax.sql.DataSource;
 
@@ -45,8 +44,6 @@ import org.springframework.transaction.support.AbstractPlatformTransactionManage
 import org.springframework.transaction.support.DefaultTransactionStatus;
 import org.springframework.transaction.support.ResourceTransactionManager;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
-import org.springframework.util.ClassUtils;
-import org.springframework.util.ReflectionUtils;
 
 /**
  * {@link org.springframework.transaction.PlatformTransactionManager}
@@ -104,15 +101,6 @@ import org.springframework.util.ReflectionUtils;
 @SuppressWarnings("serial")
 public class HibernateTransactionManager extends AbstractPlatformTransactionManager
 		implements ResourceTransactionManager, InitializingBean {
-
-	/**
-	 * A Method handle for the <code>SessionFactory.getCurrentSession()</code> method.
-	 * The return value differs between Hibernate 3.x and 4.x; for cross-compilation purposes,
-	 * we have to use reflection here as long as we keep compiling against Hibernate 3.x jars.
-	 */
-	private static final Method getCurrentSessionMethod =
-			ClassUtils.getMethod(SessionFactory.class, "getCurrentSession");
-
 
 	private SessionFactory sessionFactory;
 
@@ -293,7 +281,7 @@ public class HibernateTransactionManager extends AbstractPlatformTransactionMana
 		}
 		else if (this.hibernateManagedSession) {
 			try {
-				Session session = (Session) ReflectionUtils.invokeMethod(getCurrentSessionMethod, this.sessionFactory);
+				Session session = this.sessionFactory.getCurrentSession();
 				if (logger.isDebugEnabled()) {
 					logger.debug("Found Hibernate-managed Session [" + session + "] for Spring-managed transaction");
 				}
@@ -337,7 +325,7 @@ public class HibernateTransactionManager extends AbstractPlatformTransactionMana
 
 		try {
 			if (txObject.getSessionHolder() == null || txObject.getSessionHolder().isSynchronizedWithTransaction()) {
-				Session newSession = SessionFactoryUtils.openSession(getSessionFactory());
+				Session newSession = getSessionFactory().openSession();
 				if (logger.isDebugEnabled()) {
 					logger.debug("Opened new Session [" + newSession + "] for Hibernate transaction");
 				}
