+++ spring-orm-hibernate4/src/main/java/org/springframework/orm/hibernate4/package-info.java
