+++ spring-orm-hibernate4/src/main/java/org/springframework/orm/hibernate4/support/package-info.java
@@ -1,6 +1,6 @@
 
 /*
- * Copyright 2002-2011 the original author or authors.
+ * Copyright 2002-2012 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
